merge Feature/data input vào dev

+++ BE/Models/KetQuaHocTap.cs
@@ -6,7 +6,7 @@ public class KetQuaHocTap
     public string MaSV { get; set; } = null!;
     public int HocKy { get; set; }
     public string NamHoc { get; set; } = null!;
-    public double GPA { get; set; }
+    public float GPA { get; set; }
     public int SoTC { get; set; }
     public int? MaCB_Nhap { get; set; }
